+++ ICCardManager/docs/design/05_クラス設計書.md
@@ -241,7 +241,7 @@ graph TB
 | LedgerDetailChronologicalSorter | 利用詳細の時系列ソート・残高再計算 |
 | OperationLogger | 操作ログの記録 |
 | CsvExportService | CSV形式でのデータエクスポート |
-| CsvImportService | CSV形式でのデータインポート |
+| CsvImportService | CSV形式でのデータインポート（partial classで責務分割: Card/Staff/Ledger/Detail） |
 | PrintService | 帳票の印刷・FlowDocument生成 |
 | InputSanitizer | 入力値のサニタイズ処理（静的クラス） |
 | StationMasterService | 駅コードから駅名への解決（シングルトン） |
